foomatic-rip: Fixed reading PPD file when name is supplied by environment variable.

+++ filter/foomatic-rip/foomaticrip.c
@@ -931,8 +931,11 @@ int main(int argc, char** argv)
        command line and the options */
     if (spooler == SPOOLER_CUPS && job->printer && strlen(job->printer) > 0) {
       str = cupsGetPPD(job->printer);
-      read_ppd_file(str);
-      unlink(str);
+      if (str) {
+        read_ppd_file(str);
+        unlink(str);
+      } else
+        read_ppd_file(job->ppdfile);
     } else 
       read_ppd_file(job->ppdfile);
